Phone number field made BigInteger due to overflow

+++ Backend/LogicPersistence/LogicPersistence.Api/Models/DTOs/VictimDtos.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel.DataAnnotations;
+using System.Numerics;
 
 namespace LogicPersistence.Api.Models.DTOs;
 
@@ -26,7 +27,7 @@ public class VictimCreateDto {
 
 	[Required]
 	[Range(100000000, 99999999999)]
-	public int phone_number { get; set; }
+	public BigInteger phone_number { get; set; }
 
 	[Required]
 	[MaxLength(255)]
@@ -66,7 +67,7 @@ public class VictimUpdateDto {
 
 	[Required]
 	[Range(100000000, 99999999999)]
-	public int phone_number { get; set; }
+	public BigInteger phone_number { get; set; }
 
 	[Required]
 	[MaxLength(255)]
@@ -86,7 +87,7 @@ public class VictimDisplayDto {
 	public string name { get; set; } = string.Empty;
 	public string surname { get; set; } = string.Empty;
 	public int prefix { get; set; }
-	public int phone_number { get; set; }
+	public BigInteger phone_number { get; set; }
 	public string address { get; set; } = string.Empty;
 	public string identification { get; set; } = string.Empty;
